refactor: move bridge_client into bridge::client (namespace symmetry)
Top-level `src/bridge_client/` is conceptually the WS-client half of
the bridge subsystem; `src/bridge/serve.rs` is the MCP HTTP-server half
the agents talk to. Surface them as siblings under `bridge`:

  before:  src/bridge_client/{mod,ws,reconcile,local_store}.rs
           pub mod bridge_client (in lib.rs)
           chorus::bridge_client::run_bridge_client(...)

  after:   src/bridge/client/{mod,ws,reconcile,local_store}.rs
           pub mod client (in src/bridge/mod.rs)
           chorus::bridge::client::run_bridge_client(...)

Symmetric with `chorus::bridge::serve::run_bridge_server`. Removes a
top-level module from the public crate surface and groups all bridge
machinery in one place. No code logic changed; `git mv` preserves
history. Updated:

  - lib.rs: drop `pub mod bridge_client`
  - src/bridge/mod.rs: add `pub mod client`
  - src/cli/bridge.rs: import path → `chorus::bridge::client`
  - docs/BACKEND.md: module-map paths updated to `src/bridge/client/`

Verification: 382 lib + 16 bridge_ws + 80 e2e/realtime/server tests,
clippy `-D warnings` clean, fmt clean.
