+++ src/bridge/client/local_store.rs
