+++ src/lib.rs
@@ -1,6 +1,5 @@
 pub mod agent;
 pub mod bridge;
-pub mod bridge_client;
 pub mod config;
 pub mod logging;
 pub mod server;
